Odoo/requirements.txt
streamlit>=1.24.0
pandas
python-dotenv
Pillow             # for profile image handling
watchdog           # for fast reload during development
